requirements.txt
backports.weakref==1.0.post1
bleach==1.5.0
boto==2.49.0
boto3==1.9.38
botocore==1.12.38
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gensim==3.6.0
html5lib==0.9999999
idna==2.7
jmespath==0.9.3
Markdown==3.0.1
mock==2.0.0
numpy==1.15.4
pbr==5.1.1
protobuf==3.6.1
python-dateutil==2.7.5
requests==2.20.0
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.7.1
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
urllib3==1.24.1
Werkzeug==0.14.1
